email-scripts/requirements.txt
#requirement placeholder
kafka-python==1.4.4
